+++ lib/localize/validity/t.ex
@@ -147,7 +147,7 @@ defmodule Localize.Validity.T do
 
   defp valid(key, value) when key in @valid_keys do
     case Integer.parse(value) do
-      {_integer, ""} -> {:ok, make_date_tuple(value)}
+      {_integer, ""} when byte_size(value) in [4, 6, 8] -> {:ok, make_date_tuple(value)}
       _other -> {:error, U.invalid_value_error(key, value)}
     end
   end